Complete implementation of matching; move generic parts to Image class
